Week6 (26-30)/Day_29/backups/GenAI Powered E-Commerce Project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
orjson==3.10.7
openai==1.45.0
chromadb>=1.5.1
pymongo>=4.6.0
pandas>=2.0.0
openpyxl>=3.1.0
httpx==0.27.2